requirements.txt
accelerate==1.11.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
Authlib==1.6.5
bitsandbytes==0.48.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.2
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.6.0
filelock==3.20.0
fire==0.7.1
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.9.0
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.11.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
networkx==3.5
numpy==2.2.6
openai==2.5.0
opencv-python-headless==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pdf2docx==0.5.8
#peft==0.17.1
pillow==11.3.0
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.2
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
Pygments==2.19.2
PyMuPDF==1.26.7
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
safetensors==0.6.2
sentry-sdk==2.45.0
setuptools==70.2.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
termcolor==3.2.0
tiktoken==0.12.0
tokenizers==0.22.1
#torchvision
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
validators==0.35.0
watchfiles==1.1.1
weaviate-client==4.18.1
websockets==15.0.1
xai-sdk==1.3.1
yarl==1.22.0
zipp==3.23.0
